+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to track changes made to the project over time.
 
+## [Unreleased]
+### Added
+- Added `append_body_to_joint` method to the model to append add inertias to joints with respect to its placement.
+
 ## [0.0.2] - 2026-02-08
 ### Added
 ### Algorithms